Add files via upload
This code is a Python script that uses the Kivy library to build a graphical user interface (GUI) for our application wich is called "ASL Translator". The application uses a pre-trained computer vision model to perform object detection on video frames captured from a webcam, and displays the resulting frames with overlaid bounding boxes and labels on the GUI.
The script begins by importing various modules and libraries, including Kivy, NumPy, OpenCV, and PyTorch. It then defines a class called ASL_Translator, which subclasses App from Kivy. This class contains several methods and variables that are used to build and update the GUI.
The build method is used to build and layout the initial GUI, which consists of an Image widget to display the video frames, a Button widget to start and stop the translation, and a Label widget to display the translated text. The update method is scheduled to run at a regular interval (30 times per second) using the Clock module from Kivy. This method captures video frames from the webcam using OpenCV, passes the frames through the pre-trained object detection model, and updates the Image widget with the resulting frames.
Finally, the script calls the run method on an instance of ASL_Translator to start the application.