Rename KivyApp_to_APK.ipynb to Kivy_Android_App_Dependency_Setup.ipynb
Various libraries and dependencies required for building an Android app using the Kivy framework are installed by this code. The buildozer library, which is a tool used to package Python applications for Android, is installed, as well as cython, which is a Python compiler.

A number of other libraries and tools required for building and running Kivy applications on Android are also installed, including sdl2, sdl2-image, sdl2-mixer, sdl2-ttf, portmidi, swscale, avformat, and avcodec. Build dependencies such as build-essential, git, and python3-dev are also installed.

Libraries required for playing audio and video in Kivy applications are installed, including gstreamer and its associated plugins